+++ Gui/MerchantFrame.lua
@@ -30,8 +30,8 @@ do -- [[ Set some permanent MerchantFrame changes ]]
 	MerchantPrevPageButton:SetPoint("BOTTOMLEFT", MerchantFrameBottomLeftBorder, "TOPLEFT", 8, -5);
 	MerchantNextPageButton:SetPoint("BOTTOMRIGHT", KrowiEVU_BottomExtensionRightBorder, "TOPRIGHT", -7, -5);
 
-	-- MerchantFrameLootFilter:Hide();
-	MerchantFrameLootFilter:SetPoint("TOPRIGHT", MerchantFrame, -150, -28);
+	MerchantFrameLootFilter:Hide();
+	-- MerchantFrameLootFilter:SetPoint("TOPRIGHT", MerchantFrame, -150, -28);
 end
 
 hooksecurefunc("MerchantFrame_UpdateMerchantInfo", function()
@@ -48,9 +48,19 @@ hooksecurefunc("MerchantFrame_UpdateMerchantInfo", function()
 	if numExtraColumns > 0 then
 		KrowiEVU_BottomExtensionLeftBorder:Show();
 		KrowiEVU_BottomExtensionMidBorder:Show();
+		-- Re-arange Filter Button and Search Box if extra columns are shown
+		KrowiEVU_FilterButton:ClearAllPoints();
+		KrowiEVU_FilterButton:SetPoint("TOPRIGHT", -10, -31);
+		KrowiEVU_SearchBox:ClearAllPoints();
+		KrowiEVU_SearchBox:SetPoint("RIGHT", KrowiEVU_FilterButton, "LEFT", -10, 0);
 	else
 		KrowiEVU_BottomExtensionLeftBorder:Hide();
 		KrowiEVU_BottomExtensionMidBorder:Hide();
+		-- Re-arange Filter Button and Search Box if no extra columns are shown
+		KrowiEVU_FilterButton:ClearAllPoints();
+		KrowiEVU_FilterButton:SetPoint("TOPRIGHT", -10, -21);
+		KrowiEVU_SearchBox:ClearAllPoints();
+		KrowiEVU_SearchBox:SetPoint("TOPRIGHT", KrowiEVU_FilterButton, "BOTTOMRIGHT", 0, 2);
 	end
 	KrowiEVU_BottomExtensionRightBorder:Show();
 end);
@@ -62,13 +72,13 @@ hooksecurefunc("MerchantFrame_UpdateBuybackInfo", function()
 	KrowiEVU_BottomExtensionRightBorder:Hide();
 end);
 
-addon.CachedItemIndices = {};
-
 hooksecurefunc("MerchantFrame_UpdateRepairButtons", function()
 	if not CanMerchantRepair() then
 		MerchantSellAllJunkButton:SetPoint("BOTTOMRIGHT", MerchantFrame, "BOTTOMLEFT", 162, 33);
 	end
 end);
+
+addon.CachedItemIndices = {};
 
 local origGetMerchantItemInfo = GetMerchantItemInfo;
 GetMerchantItemInfo = function(index)